register providers for testing
In order to run some feature tests, it was necessary to adapt the  code,
including registering some providers for the tests, such as the  sanctum
provider and the route provider.

This is what this commit does:  adapt  the  tests  to  work  within  the
package structure.

The source code contains several classes, many of which are used only in
tests due. The user will end up using only Trait classes, which is  what
this package is intended for. Nonetheless, in  order  to  ensure  proper
program behavior, it is necessary to simulate the application as  if  it
were being run  by  an  actual  user  using  this  package.  This  means
simulating authorizaton handling of http requests, which requires set up
of routes, controllers, and a database.