remove unnecessary contracts
Some contracts are unnecessary, such as the  HasRoleContract.  The  user
should be able to use a package trait without having to also  explicitly
declare that the class is implementing the contracts for that trait. The
only exception to this is the ResourcePolicy contract, because the  user
does need to implement  a  method  to  use  the  trait.  All  the  other
contracts are verbose nonsense that harms user experience.