add trait that provides Permissionable and HasPermission behavior
Add a trait that provides the behavior of both mentionend  traits.  This
is due to a method collision, which prevented a class  from  using  both
traits without having to explicitly specify from which trait to use  the
conflicted method from. Moreover, it is easier from a  user  perspective
to import only a single trait than to import two.